requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
psycopg2-binary

pyyaml
scikit-learn

# openmeteo-requests
requests-cache
retry-requests

## Too large package size
# pandas
# catboost
# joblib